refactor: Dao 와 DB connect 로직을 분리
