+++ src/main/java/chess/dao/ChessGameDao.java
@@ -1,40 +1,23 @@
 package chess.dao;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+import chess.DBConnector;
 import chess.domain.ChessGame;
 
 public class ChessGameDao {
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
+    private final DBConnector dbConnector;
 
-    private Connection getConnection() {
-        loadDriver();
-        Connection connection = null;
-        try {
-            connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return connection;
-    }
-
-    private void loadDriver() {
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        } catch (final Exception e) {
-            e.printStackTrace();
-        }
+    public ChessGameDao() {
+        this.dbConnector = new DBConnector();
     }
 
     public void save(String gameID, ChessGame chessGame) {
         String sql = "insert into chessGame (gameID, turn) values (?, ?)";
-        try (Connection connection = getConnection();
+        try (Connection connection = dbConnector.getConnection();
              PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.setString(1, gameID);
             statement.setString(2, chessGame.getTurn().name());
@@ -46,7 +29,7 @@ public class ChessGameDao {
 
     public void updateTurn(String gameID, ChessGame chessGame) {
         String sql = "update chessGame set turn = ? where gameID = ?";
-        try (Connection connection = getConnection();
+        try (Connection connection = dbConnector.getConnection();
              PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.setString(1, chessGame.getTurn().name());
             statement.setString(2, gameID);
@@ -59,7 +42,7 @@ public class ChessGameDao {
     public String findTurnByID(String gameID) {
         String sql = "select turn from chessGame where gameID = ?";
         String turn = null;
-        try (Connection connection = getConnection();
+        try (Connection connection = dbConnector.getConnection();
              PreparedStatement statement = getPreparedStatement(gameID, sql, connection);
              ResultSet resultSet = statement.executeQuery()) {
             if (!resultSet.next()) {
